extracted company data, build search route
