--- ReadMe.md
@@ -1,1 +1,0 @@
-## EquiPay
