--- models/company.js
@@ -1,14 +1,0 @@
-const { ObjectId } = require("mongodb");
-const mongoose = require("mongoose");
-const { Schema } = mongoose;
-
-const companySchema = new Schema({
-  name: { type: String, required: true },
-  reviews: [
-    {
-      type: Schema.Types.ObjectId,
-      ref: "review",
-    },
-  ],
-});
-module.exports = mongoose.model("company", companySchema);
