--- models/message.js
@@ -1,40 +1,0 @@
-const { ObjectId } = require("mongodb");
-const mongoose = require("mongoose");
-const { Schema } = mongoose;
-
-const messageSchema = new Schema(
-  {
-    userid: {
-      type: Schema.Types.ObjectId,
-      required: true,
-    },
-    message: {
-      type: String,
-      required: true,
-    },
-    likes: { type: Number, default: 0 },
-    dislikes: { type: Number, default: 0 },
-    likedBy: [
-      {
-        type: Schema.Types.ObjectId,
-        ref: "user",
-      },
-    ],
-    dislikedBy: [
-      {
-        type: Schema.Types.ObjectId,
-        ref: "user",
-      },
-    ],
-    comments: [
-      {
-        type: Schema.Types.ObjectId,
-        ref: "comment",
-      },
-    ],
-  },
-  {
-    timestamps: true,
-  }
-);
-module.exports = mongoose.model("message", messageSchema);
