--- models/user.js
@@ -1,24 +1,0 @@
-const { ObjectId } = require("mongodb");
-const mongoose = require("mongoose");
-const { Schema } = mongoose;
-
-const userSchema = new Schema(
-  {
-    name: { type: String, required: true },
-    age: { type: Number, required: true },
-    gender: { type: String, required: true },
-    email: { type: String, required: true },
-    username: String,
-    password: String,
-    reviews: [
-      {
-        type: Schema.Types.ObjectId,
-        ref: "review",
-      },
-    ],
-  },
-  {
-    timestamps: true,
-  }
-);
-module.exports = mongoose.model("user", userSchema);
