--- package.json
@@ -1,25 +1,0 @@
-{
-  "name": "hackprojectn",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "bcrypt": "^5.1.1",
-    "body-parser": "^1.20.2",
-    "connect-flash": "^0.1.1",
-    "connect-mongo": "^5.0.0",
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "express-session": "^1.17.3",
-    "flash": "^1.1.0",
-    "mongoose": "^7.5.2",
-    "passport": "^0.6.0",
-    "passport-local": "^1.0.0"
-  }
-}
