+++ server/ReadMe.md
@@ -1,0 +1,1 @@
+## EquiPay
