+++ server/company_List_Extractor/.gitignore
@@ -1,0 +1,2 @@
+/companies
+companies.json
